Resolve skypilot.py conflict where both branches independently solved
emitting a long-running SERVICE's URL binding before terminal status:

- HEAD: per-poll log scrape with offset tracking + static-binding backoff
- upstream/main: live log streaming via sky.tail_logs(follow=True)

Adopt upstream's live-stream mechanism (lower latency, lower I/O, no
static-binding prediction) and drop the per-poll scrape. Keep this
branch's launch_skypilot_teardown (SERVICE teardown on LSF) and upstream's
provision-retry helpers, _skypilot_retry_in_progress_events, and
_relaunch_attempts cleanup. Remove the now-obsolete
test_skypilot_service_binding.py, which only exercised the dropped
per-poll scrape.

Verified: 170 environment unit tests pass.

+++ src/gbserver/buildrunner/abstractbuildrunner.py
@@ -51,8 +51,10 @@ class AbstractBuildRunner(ABC):
     # The git token to use to update the PR of the build, if any.
     gh_token: str = GBSERVER_GITHUB_TOKEN
 
-    # Looping interval for worker function
-    monitoring_interval: int
+    # Looping interval for worker function (seconds). Bounded at construction by
+    # the watcher configs (Field ge=MIN_MONITORING_INTERVAL_SECONDS) and the
+    # build-runner CLI (IntRange), so a busy-looping 0 never reaches here.
+    monitoring_interval: float
 
     # Git api endpoint to register messages in the PR
     gh_api_endpoint: str
